+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.6.0 2020-07-12
+
+Added IAM stateless module. 
+
+Accepts `list`s of groups, users and policies that can be used to create these resources and connect them by names references.
+
+e.g. `var.users.*.group_names` determines to wich groups the user will be attached to, but will not create the group. 
+That must be done with `var.groups`.
+
 ## v0.5.0 2020-07-12
 
 Changed AWS Organizations module to be stateless, removed hardcoded account and organizational units logic.